Move common tools files into separeate lib dir
Save duplication of snapserve and catalyst files by creating
a common dir and moving a bunch of files in there.

Add tools/CMakeLists so we can build catalyst and snapserve in one pass
with a single cmake build.

Fix a couple of std:: problems when we need to link with stdlibc++